Debug Version 0.5.1 stable
Moving GIT repository from evandenny to EvDenny for ease of access. May delete repository on evandenny. Currently on Debug Version 0.5.1 of AppCPP.
Recent changes include an overhaul of the menu system to incorporate the functionality of menu.h that I set up to the main file App.cpp. The old menu system was deprecated and I began moving to a simplified, easier to understand way of moving through the menus. A series of enumerated boolean expressions would determine what would appear on the screen next.
For example, at the beginning of the Main Menu, the value of each of the 5 menus is reset to the ::Null state. This makes the program stop at the first user input state where the user can choose one of the options: Play Games, Tutorials, or Exit. Each of the options will change the value of an object instance of MainMenu:: called menuOption.
menuOption defaults to ::null. if the user chooses something other than 1, 2,or 3, the program ignores the users answer, sets the object to ::null, and redisplays the menu. If one of the options is selected, the value of menuOption changes and allows the user to pass to the menu of their choosing. Play Games displays the gameMenu, Tutorials displays the tutorMenu, and Exit exits the program with Menu::exiting();
The system is more complicated but works much faster than previous versions of AppCPP.