Debug Version 0.5.1_15
Recreated mdy.cpp as mdy2.cpp for Windows.
Used class to create Calendar class object that functions as a display for current date and time (Not fully implemented in object yet).
Current workaround is creating a tm struct in the main() function and creating individual int objects for each piece of the datetime strings.
void hms(int value) is used to add an extra '0' ahead of the hour, minute, or second ints when they have values less than 10.
*** 5:1:3 is converted to 05:01:03 for readability.
Calendar object is constructed as cal.
*** Constructor function sets boolean values of 'future', 'past', '_continue', and 'finish' to false. int _count is set to 0. int Today is set to current dotw in int form (0 to 6). mainloop (forward or backward) manipulates each of these variables after the user inputs a valid date. Once the current date is == the input date by advancing or the day forward/backward by 1, advancing the Today int forward/backward by 1 (and setting the var == 0 or 6 whenever it exits the 0-6 range to rotate through the days of week alongside the date) increasing the count, and changing the month and year when the  day is outside the valid calendar date range, the _count variable is displayed along with the Today variable (converted to a string (0 = 'Sunday', 3 = 'Wednesday').
This works similar to the mdy.cpp executable, but without as much obfuscation. The steps are simplified and less convoluted in the way that functions in mdy.cpp were nested and nested continuously within each other, making debugging and even reading the code very difficult.
mdy2.cpp has been relatively easy to debug in comparison.
For example, the mainloop was bugged and would count to infinity because the order of functions on the counting function was bracketed wrong. It was easy to fix and spot.
Unlike mdy.cpp, this version seems to run better and faster in comparison. I believe this is due to my use of inline functions whenever I could within the Calendar class to make the compiler easily use simple functions rather than digging through memory to find each one.
Currently the program is working as intended but more bug testing needs to be done.
I know of one very specific bug, that should be simple to fix:
BUG->Lead days like 2-29-2024 are not counted or even accessible by the Calendar's mainloop functions.
getInput() can input the date as intended, but when the program runs its mainloop, the count reaches 2-28-2024, and then runs nextMonth() setting the next day to 3-1-2024 and increasing the count, before running check().
This means every leap day is skipped.
So inputing 1-1-10000, will be hundreds of days off.
Gregorian Calendars also added a day every 100 or 400 years on dates that give 0 to value%100/400 I believe, so additional checks for years like 2400 may need to be implemented to account for the precession of the Earth. (Thanks Einstein...)